weather-etl-project/requirements.txt
requests>=2.28.0
pandas>=1.5.0
psycopg2-binary>=2.9.0
apscheduler>=3.10.0
